JDBC connection is working. CRUD operations for Users (Patient, Doctor) are working.
Heavily modified constructor for User. Now, instead of ID's, we will require DNI. DNI is now the Primary Key in the Database and it is no longer auto-generated.

Email has been set to UNIQUE field in the Database. This correlates with the way of logging in: via Email and Password.

Interface now needs to be modified in order to retrieve DNI from users upon registering.