Added a layer of security to credentials for DB Driver (imports them from local environment variables).
JavaMailUtilty class created and finished. Needs to be thoroughly tested.

Modified Patient to be able to generate a random verification code upon creation. Also added a new constructor to be able to retrieve the Patient object from the DB. Modified implementations of Patient for DB to adjust to the verification code addition.

Started working on sending verification code via interface.